Added SWEET32 and TLS1 script
Updated Version 3.0

The previous version had a bug (for a mistyping error) that blocked the RDP access for an incomplete sentence. The Fix in that typo error has been applied and now it's working the way it should. I've tested in my JCS's environments without issues..

You can use the Get-Help .\sweet32.ps1 -detailed to get help in console.

 



I want to thank  @bsjj2727 for finding the bug and also reporting the solution in here. Thank you So much!

 

Updated Version 2.0

The previous version has a bug that was solved with this update, if you downloaded the previous version, just rerun the script with the same options.

 

Version 1.0 Description

This script will solve two vulnerabilities, SWEET32 Birthday attack and TLS1.0, required for a PCI Compliance in your web server infrastructure, and this includes Exchange Server and Windows servers from 2008 up to our days. (2016).

 

How to use the script:

As any other script you would need to modify the executionpolicy, by running in an elevated PowerShell Console:

 

PowerShell
Set-ExecutionPolicy Unrestricted and accepting the message