Create New Linux Machine for TS3
his script will create a whole infrastructure to create a single server in windows azure with a static public IP. The idea of this script is to use a Linux server as TeamSpeak 3 server. 

 

TeamSpeak 3 server offers "the ideal tool for education, training, online gaming, internal business communication, and staying in touch with friends and family...". "It's capable of handling thousands of simultaneous users..." TeamSpeak's team.



https://gallery.technet.microsoft.com/scriptcenter/Create-Linux-VM-for-62886707
Basically, I have worked with windows server most of my time so I wanted to publish something different using Linux and exploring the gamer side.

This script working can be seen on youtube in this link: https://youtu.be/CVLODajWeic (Corrected URL).